requirements.txt
pandas==1.3.3
google-cloud-storage==1.44.0
Flask==2.0.1

